Solana client should be replaced.
