+++ gold-bot/Cargo.toml
@@ -4,12 +4,10 @@ version = "0.0.0-alpha"
 edition = "2021"
 
 [dependencies]
-agsol-gold-client = { path = "../client" }
 agsol-gold-contract = { path = "../contract", features = ["client"] }
 anyhow = "1"
 env_logger = "0.9.0"
 log = "0.4"
 solana-sdk = "1.9.0"
-solana-client = "1.9.0"
+#solana-client = "1.9.0"
 structopt = "0.3"
-
